优化sync

+++ sync.sh
@@ -3,7 +3,7 @@ trap 'cp /jd-scripts-docker/sync.sh /sync' Exit
 git clone --depth=1 https://github.com/jw10126121/jd-scripts-docker.git /jd-scripts-docker_tmp
 [[ -d /jd-scripts-docker_tmp ]] && {
   echo '更新jd-scripts-docker'
-  rm -rf /jd-scripts-docker
+  [[ -d /jd-scripts-docker ]] && rm -rf /jd-scripts-docker
   mv /jd-scripts-docker_tmp /jd-scripts-docker
   [[ ! -e /codeKey ]] && cat /jd-scripts-docker/env/codeKey > /codeKey
 }
@@ -33,8 +33,9 @@ git clone --depth=1 https://github.com/jw10126121/jd-scripts-docker.git /jd-scri
 
 cd /scripts || exit 1
 npm install || npm install --registry=https://registry.npm.taobao.org || exit 1
-[[ -e /crontab.list ]] && cp /crontab.list /crontab.list.old
-[[ -e /crontab.list ]] && rm -fr /crontab.list
+
+# 备份crontab.list
+[[ -e /crontab.list ]] && mv /crontab.list /crontab.list.old
 
 echo '更新crontab'
 cp /jd-scripts-docker/crontab.list /crontab.list
